Moved search componenents to own directory.
